Update dealing with breaking changes to ENA
Changes are twofold
1. ENA will cease serving XML, so instead we need to download the flatfile and convert it to XML locally using the ENA API
2. ENA will cease serving individual WGS entries, so instead we need to download and convert the whole WGS set. This isn't as bad as it sounds as we'll probably need all the entries locally.
To deal with these, the code for downloading and converting ENA data before parsing is now contained in EnaXmlRetriever (woof) which is passed to the materializer instead of the executor (the executor is now used internally by the retriever and not exposed to the materializer directly). This has affected a large number of processors who now have different constructor signatures.
Lastly, due to ENA redirecting https to compressed FTP, the Apache httpclient is used instead of the simple JRE URL connection functionality.